requirements.txt
instaloader==4.13
pydantic-settings
requests
python-telegram-bot==21.4
tenacity
